Change some missed doubles to longs

+++ src/main/java/moze_intel/projecte/gameObjs/gui/GUICollectorMK3.java
@@ -37,8 +37,8 @@ public class GUICollectorMK3 extends GuiContainer
 	protected void drawGuiContainerForegroundLayer(int var1, int var2)
 	{
 		this.fontRenderer.drawString(Long.toString(container.emc), 91, 32, 4210752);
-		
-		double kleinCharge = container.kleinEmc;
+
+		int kleinCharge = container.kleinEmc;
 		if (kleinCharge > 0)
 			this.fontRenderer.drawString(Constants.EMC_FORMATTER.format(kleinCharge), 91, 44, 4210752);
 	}
